pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "openai>=1.86.0",
    "pandas>=2.3.0",
    "plotly>=6.1.2",
    "psycopg2-binary>=2.9.10",
    "streamlit>=1.45.1",
]
